Add signing function on vote
