+++ public/assets/css/component-style.css
@@ -120,6 +120,9 @@
 }
 
 .nav-btn {
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
   position: relative;
   top: 10px;
 }
